+++ src/main/java/pl/connectis/cinemareservationsapp/controller/RoomController.java
@@ -1,42 +1,77 @@
 package pl.connectis.cinemareservationsapp.controller;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import pl.connectis.cinemareservationsapp.exceptions.BadRequestException;
+import pl.connectis.cinemareservationsapp.exceptions.ResourceNotFoundException;
 import pl.connectis.cinemareservationsapp.model.Room;
 import pl.connectis.cinemareservationsapp.service.RoomService;
 
 import javax.validation.Valid;
-import java.util.List;
 
 @RestController
+@RequestMapping("/room")
 public class RoomController {
 
     @Autowired
     private RoomService roomService;
 
-    @GetMapping("/room/all")
+    @GetMapping("/all")
     public Iterable<Room> getAllRooms() {
         return roomService.findAll();
     }
 
-    @GetMapping("/room/{id}")
-    public List<Room> getRoomById(@PathVariable long id) {
-        return roomService.findById(id);
+    @GetMapping
+    public Room getRoomById(@RequestParam long id) {
+        Room room = roomService.findById(id);
+        if(room == null) {
+            throw new ResourceNotFoundException("room {id=" + id + "} was not found");
+        }
+        return room;
     }
 
-    @PostMapping("/room")
-    public Room addRoom(@Valid @RequestBody Room room) {
-        return roomService.save(room);
+    @PostMapping
+    public ResponseEntity<Room> addRoom(@Valid @RequestBody Room room) {
+        validateRoom(room);
+        return new ResponseEntity<>(roomService.save(room), HttpStatus.CREATED);
     }
 
-    @PostMapping("/room/many")
-    public Iterable<Room> addRoomList(@Valid @RequestBody Iterable<Room> roomList) {
-        return roomService.saveAll(roomList);
+    @PostMapping("/many")
+    public ResponseEntity<Iterable> addRoomList(@Valid @RequestBody Iterable<Room> roomList) {
+        for (Room room : roomList) {
+            validateRoom(room);
+        }
+        return new ResponseEntity<>(roomService.saveAll(roomList), HttpStatus.CREATED);
     }
 
-    @DeleteMapping("/room/{id}")
-    public void deleteRoom(@PathVariable long id) {
+    @PutMapping
+    public ResponseEntity<Room> updateById(@RequestParam long id, @Valid @RequestBody Room room) {
+        Room existingRoom = roomService.findById(id);
+        if(existingRoom == null) {
+            throw new ResourceNotFoundException("room {id=" + id + "} was not found");
+        }
+        validateRoom(room);
+        return new ResponseEntity<>(roomService.updateById(id, room), HttpStatus.CREATED);
+    }
+
+    @DeleteMapping
+    public ResponseEntity deleteRoom(@RequestParam long id) {
+        Room room = roomService.findById(id);
+        if (room == null) {
+            throw new ResourceNotFoundException("room {id=" + id + "} was not found");
+        }
         roomService.deleteById(id);
+        return new ResponseEntity(HttpStatus.NO_CONTENT);
+    }
+
+    private void validateRoom(Room room) {
+        if (!roomService.validateCapacity(room)) {
+            throw new BadRequestException("capacity does not correspond to layout");
+        } else if (room.getCapacity() == 0) {
+            roomService.setCapacityFromLayout(room);
+        }
     }
 
 }
